+++ src/extension.ts
@@ -12,6 +12,9 @@ import * as RPC from "./modules/rpc.mjs";
 // @ts-ignore
 import config from "./config.json" assert { type: `json` };
 
+// Commands
+import { getCommands } from "./modules/commands";
+
 interface LooseObject {
     [key: string]: any
 }
@@ -23,33 +26,25 @@ export function activate({ subscriptions }: vscode.ExtensionContext): void {
 	// Config setup
 	config.extension.startTimestamp = new Date();
 	config.extension.settings.rpc = getSettings();
-
-	RPC.init({ clientId: config.application.id, intervalTime: config.extension.settings.rpc.updateTimeInterval * 1000 });
-	
-	// Adding a command "reloadRPC"
-	const reloadCommand = "vscode-discord-rpc.reloadRPC";
-	subscriptions.push(vscode.commands.registerCommand(reloadCommand, () => {
-		config.extension.settings.rpc = getSettings();
-
-		RPC.reload({ intervalTime: config.extension.settings.rpc.updateTimeInterval * 1000 });
-	}));
-
-	const stopCommand = "vscode-discord-rpc.stopRPC";
-	subscriptions.push(vscode.commands.registerCommand(stopCommand, () => RPC.stop()));
-
-	// On file open event
-	subscriptions.push(vscode.window.onDidChangeActiveTextEditor(() => RPC.update()));
 	
 	// Adding a statusBarItem to status bar with the command "reloadRPC" and loading text
 	let statusBarItem = vscode.window.createStatusBarItem(vscode.StatusBarAlignment.Left, 0);
 
+	// Registering commands
+	const commands = registerCommands(subscriptions, statusBarItem);
+
+	// On file open event
+	subscriptions.push(vscode.window.onDidChangeActiveTextEditor(() => RPC.update()));
+
 	statusBarItem.text = "$(sync~spin) RPC Connecting...";
-	statusBarItem.command = reloadCommand;
+	statusBarItem.command = commands.barItem;
 
 	subscriptions.push(statusBarItem);
 	statusBarItem.show();
 
 	registerRpcEvents(statusBarItem);
+
+	RPC.init({ clientId: config.application.id, intervalTime: config.extension.settings.rpc.updateTimeInterval * 1000 });
 }
 
 // The extension has stopped
@@ -104,6 +99,23 @@ function getActivity(editor: vscode.TextEditor | undefined, config: any): Object
 	return activity;
 }
 
+function registerCommands(subscriptions: any, statusBarItem: vscode.StatusBarItem) {
+	// Adding a command "startRPC"
+	const commands = getCommands(config, statusBarItem);
+
+	let parsedCommands: LooseObject = {};
+
+	for (const command of commands) {
+		const fullName = `vscode-discord-rpc.${command.name}`;
+
+		subscriptions.push(vscode.commands.registerCommand(fullName, command.execute));
+
+		parsedCommands[command.name] = fullName;
+	}
+	
+	return parsedCommands;
+}
+
 function registerRpcEvents(statusBarItem: vscode.StatusBarItem): void {
 	RPC.on(RPC.Events.Ready, (rpc: any) => {
 		console.log(`${rpc.user.username} connected!`);
@@ -121,8 +133,6 @@ function registerRpcEvents(statusBarItem: vscode.StatusBarItem): void {
 
 	RPC.on(RPC.Events.Update, (rpc: any) => {
 		console.log("Updated RPC");
-
-		config.extension.settings.rpc = getSettings();
 
 		rpc.setActivity(getActivity(vscode.window.activeTextEditor, config));
 	});